# Conflicts:
#	src/Interpreters/DDLTask.cpp

+++ src/Disks/ObjectStorages/DiskObjectStorage.cpp
@@ -11,7 +11,6 @@
 #include <Common/filesystemHelpers.h>
 #include <Common/CurrentMetrics.h>
 #include <Common/Scheduler/IResourceManager.h>
-#include <Disks/ObjectStorages/DiskObjectStorageRemoteMetadataRestoreHelper.h>
 #include <IO/CachedInMemoryReadBufferFromFile.h>
 #include <Disks/IO/ReadBufferFromRemoteFSGather.h>
 #include <Disks/IO/AsynchronousBoundedReadBuffer.h>
@@ -28,6 +27,8 @@ namespace DB
 namespace ErrorCodes
 {
     extern const int INCORRECT_DISK_INDEX;
+    extern const int LOGICAL_ERROR;
+    extern const int CANNOT_RMDIR;
 }
 
 
@@ -45,18 +46,16 @@ DiskTransactionPtr DiskObjectStorage::createObjectStorageTransaction()
 {
     return std::make_shared<DiskObjectStorageTransaction>(
         *object_storage,
-        *metadata_storage,
-        send_metadata ? metadata_helper.get() : nullptr);
+        *metadata_storage);
 }
 
-DiskTransactionPtr DiskObjectStorage::createObjectStorageTransactionToAnotherDisk(DiskObjectStorage& to_disk)
+DiskTransactionPtr DiskObjectStorage::createObjectStorageTransactionToAnotherDisk(DiskObjectStorage & to_disk)
 {
     return std::make_shared<MultipleDisksObjectStorageTransaction>(
         *object_storage,
         *metadata_storage,
         *to_disk.getObjectStorage(),
-        *to_disk.getMetadataStorage(),
-        send_metadata ? metadata_helper.get() : nullptr);
+        *to_disk.getMetadataStorage());
 }
 
 
@@ -72,10 +71,8 @@ DiskObjectStorage::DiskObjectStorage(
     , log(getLogger("DiskObjectStorage(" + name + ")"))
     , metadata_storage(std::move(metadata_storage_))
     , object_storage(std::move(object_storage_))
-    , send_metadata(config.getBool(config_prefix + ".send_metadata", false))
     , read_resource_name_from_config(config.getString(config_prefix + ".read_resource", ""))
     , write_resource_name_from_config(config.getString(config_prefix + ".write_resource", ""))
-    , metadata_helper(std::make_unique<DiskObjectStorageRemoteMetadataRestoreHelper>(this, ReadSettings{}, WriteSettings{}))
     , remove_shared_recursive_file_limit(config.getUInt64(config_prefix + ".remove_shared_recursive_file_limit", DEFAULT_REMOVE_SHARED_RECURSIVE_FILE_LIMIT))
 {
     data_source_description = DataSourceDescription{
@@ -122,16 +119,18 @@ DiskObjectStorage::DiskObjectStorage(
                             {
                                 switch (mode)
                                 {
-                                    case ASTCreateResourceQuery::AccessMode::Read: new_read_resource_name_from_sql_any.insert(resource_name); break;
-                                    case ASTCreateResourceQuery::AccessMode::Write: new_write_resource_name_from_sql_any.insert(resource_name); break;
+                                    case ResourceAccessMode::DiskRead: new_read_resource_name_from_sql_any.insert(resource_name); break;
+                                    case ResourceAccessMode::DiskWrite: new_write_resource_name_from_sql_any.insert(resource_name); break;
+                                    default: break;
                                 }
                             }
                             else if (*disk == name)
                             {
                                 switch (mode)
                                 {
-                                    case ASTCreateResourceQuery::AccessMode::Read: new_read_resource_name_from_sql.insert(resource_name); break;
-                                    case ASTCreateResourceQuery::AccessMode::Write: new_write_resource_name_from_sql.insert(resource_name); break;
+                                    case ResourceAccessMode::DiskRead: new_read_resource_name_from_sql.insert(resource_name); break;
+                                    case ResourceAccessMode::DiskWrite: new_write_resource_name_from_sql.insert(resource_name); break;
+                                    default: break;
                                 }
                             }
                         }
@@ -216,20 +215,13 @@ size_t DiskObjectStorage::getFileSize(const String & path) const
 
 void DiskObjectStorage::moveDirectory(const String & from_path, const String & to_path)
 {
-    if (send_metadata)
-        sendMoveMetadata(from_path, to_path);
-
     auto transaction = createObjectStorageTransaction();
     transaction->moveDirectory(from_path, to_path);
     transaction->commit();
 }
 
-void DiskObjectStorage::moveFile(const String & from_path, const String & to_path, bool should_send_metadata)
+void DiskObjectStorage::moveFile(const String & from_path, const String & to_path)
 {
-
-    if (should_send_metadata)
-        sendMoveMetadata(from_path, to_path);
-
     auto transaction = createObjectStorageTransaction();
     transaction->moveFile(from_path, to_path);
     transaction->commit();
@@ -267,11 +259,6 @@ void DiskObjectStorage::copyFile( /// NOLINT
     }
 }
 
-void DiskObjectStorage::moveFile(const String & from_path, const String & to_path)
-{
-    moveFile(from_path, to_path, send_metadata);
-}
-
 void DiskObjectStorage::replaceFile(const String & from_path, const String & to_path)
 {
     if (existsFile(to_path))
@@ -283,6 +270,27 @@ void DiskObjectStorage::replaceFile(const String & from_path, const String & to_
     else
         moveFile(from_path, to_path);
 }
+
+void DiskObjectStorage::renameExchange(const std::string & old_path, const std::string & new_path)
+{
+    if (existsFile(new_path))
+    {
+        auto temp_old_path = old_path + "_tmp_rename_exchange";
+        auto transaction = createObjectStorageTransaction();
+        transaction->moveFile(old_path, temp_old_path);
+        transaction->moveFile(new_path, old_path);
+        transaction->moveFile(temp_old_path, new_path);
+        transaction->commit();
+    }
+    else
+        moveFile(old_path, new_path);
+}
+
+bool DiskObjectStorage::renameExchangeIfSupported(const std::string &, const std::string &)
+{
+    return false;
+}
+
 
 void DiskObjectStorage::removeSharedFile(const String & path, bool delete_metadata_only)
 {
@@ -323,29 +331,12 @@ bool DiskObjectStorage::checkUniqueId(const String & id) const
     return object_storage->exists(object);
 }
 
-void DiskObjectStorage::createHardLink(const String & src_path, const String & dst_path, bool should_send_metadata)
+void DiskObjectStorage::createHardLink(const String & src_path, const String & dst_path)
 {
-    if (should_send_metadata && !dst_path.starts_with("shadow/"))
-    {
-        auto revision = metadata_helper->revision_counter + 1;
-        metadata_helper->revision_counter += 1;
-        const ObjectAttributes object_metadata {
-            {"src_path", src_path},
-            {"dst_path", dst_path}
-        };
-        metadata_helper->createFileOperationObject("hardlink", revision, object_metadata);
-    }
-
     auto transaction = createObjectStorageTransaction();
     transaction->createHardLink(src_path, dst_path);
     transaction->commit();
 }
-
-void DiskObjectStorage::createHardLink(const String & src_path, const String & dst_path)
-{
-    createHardLink(src_path, dst_path, send_metadata);
-}
-
 
 void DiskObjectStorage::setReadOnly(const String & path)
 {
@@ -383,17 +374,30 @@ void DiskObjectStorage::clearDirectory(const String & path)
 
 void DiskObjectStorage::removeDirectory(const String & path)
 {
+    if (!isDirectoryEmpty(path))
+        throw Exception(ErrorCodes::CANNOT_RMDIR, "Unable to remove directory '{}', the directory is not empty", path);
+
     auto transaction = createObjectStorageTransaction();
     transaction->removeDirectory(path);
     transaction->commit();
 }
 
+void DiskObjectStorage::removeDirectoryIfExists(const String & path)
+{
+    if (!existsDirectory(path))
+        return;
+    removeDirectory(path);
+}
 
 DirectoryIteratorPtr DiskObjectStorage::iterateDirectory(const String & path) const
 {
     return metadata_storage->iterateDirectory(path);
 }
 
+bool DiskObjectStorage::isDirectoryEmpty(const String & path) const
+{
+    return metadata_storage->isDirectoryEmpty(path);
+}
 
 void DiskObjectStorage::listFiles(const String & path, std::vector<String> & file_names) const
 {
@@ -440,12 +444,10 @@ void DiskObjectStorage::shutdown()
     LOG_INFO(log, "Disk {} shut down", name);
 }
 
-void DiskObjectStorage::startupImpl(ContextPtr context)
+void DiskObjectStorage::startupImpl()
 {
     LOG_INFO(log, "Starting up disk {}", name);
     object_storage->startup();
-
-    restoreMetadataIfNeeded(context->getConfigRef(), "storage_configuration.disks." + name, context);
 
     LOG_INFO(log, "Disk {} started up", name);
 }
@@ -572,15 +574,6 @@ bool DiskObjectStorage::tryReserve(UInt64 bytes)
 
     return false;
 }
-void DiskObjectStorage::sendMoveMetadata(const String & from_path, const String & to_path)
-{
-    chassert(send_metadata);
-    auto revision = metadata_helper->revision_counter + 1;
-    metadata_helper->revision_counter += 1;
-
-    const ObjectAttributes object_metadata{{"from_path", from_path}, {"to_path", to_path}};
-    metadata_helper->createFileOperationObject("rename", revision, object_metadata);
-}
 
 bool DiskObjectStorage::supportsCache() const
 {
@@ -592,6 +585,11 @@ bool DiskObjectStorage::isReadOnly() const
     return object_storage->isReadOnly();
 }
 
+bool DiskObjectStorage::isPlain() const
+{
+    return object_storage->isPlain();
+}
+
 bool DiskObjectStorage::isWriteOnce() const
 {
     return object_storage->isWriteOnce();
@@ -600,6 +598,11 @@ bool DiskObjectStorage::isWriteOnce() const
 bool DiskObjectStorage::supportsHardLinks() const
 {
     return !isWriteOnce() && !object_storage->isPlain();
+}
+
+bool DiskObjectStorage::supportsPartitionCommand(const PartitionCommand & command) const
+{
+    return !isWriteOnce() && metadata_storage->supportsPartitionCommand(command);
 }
 
 DiskObjectStoragePtr DiskObjectStorage::createDiskObjectStorage()
@@ -668,61 +671,87 @@ std::unique_ptr<ReadBufferFromFileBase> DiskObjectStorage::readFile(
     const auto storage_objects = metadata_storage->getStorageObjects(path);
     auto global_context = Context::getGlobalContextInstance();
 
-    const bool file_can_be_empty = !file_size.has_value() || *file_size == 0;
-    if (storage_objects.empty() && file_can_be_empty)
+    if (storage_objects.empty())
+    {
+        if (file_size.has_value() && *file_size != 0)
+            throw Exception(ErrorCodes::LOGICAL_ERROR,
+                "Empty list of objects for nonempty file on disk {} at {}", name, path);
         return std::make_unique<ReadBufferFromEmptyFile>();
+    }
+
+    /// Matryoshka of read buffers:
+    ///
+    /// [AsynchronousBoundedReadBuffer] (if use_async_buffer)
+    ///   [CachedInMemoryReadBufferFromFile] (if use_page_cache)
+    ///     ReadBufferFromRemoteFSGather
+    ///       [CachedOnDiskReadBufferFromFile] (if fs cache is enabled)
+    ///         ReadBufferFromS3 or similar
+    ///
+    /// Some of them have special requirements:
+    ///  * use_external_buffer = true is required for the buffer nested directly inside
+    ///    AsynchronousBoundedReadBuffer, CachedInMemoryReadBufferFromFile, and
+    ///    ReadBufferFromRemoteFSGather.
+    ///  * The buffer directly inside CachedInMemoryReadBufferFromFile must be freely seekable.
+    ///    I.e. either remote_read_buffer_restrict_seek = false or buffer implementation that
+    ///    ignores that setting.
+    ///    Note: ReadBufferFromRemoteFSGather ignores this setting.
 
     auto read_settings = updateIOSchedulingSettings(settings, getReadResourceName(), getWriteResourceName());
     /// We wrap read buffer from object storage (read_buf = object_storage->readObject())
     /// inside ReadBufferFromRemoteFSGather, so add nested buffer setting.
     read_settings = read_settings.withNestedBuffer();
 
+    const bool use_async_buffer = read_settings.remote_fs_method == RemoteFSReadMethod::threadpool;
+    const bool use_page_cache =
+        (!object_storage->supportsCache() || !read_settings.enable_filesystem_cache)
+        && read_settings.page_cache && read_settings.use_page_cache_for_disks_without_file_cache;
+
+    const bool use_external_buffer_for_gather = use_async_buffer || use_page_cache;
+
     auto read_buffer_creator =
         [this, read_settings, read_hint, file_size]
         (bool restricted_seek, const StoredObject & object_) mutable -> std::unique_ptr<ReadBufferFromFileBase>
     {
         read_settings.remote_read_buffer_restrict_seek = restricted_seek;
-        auto impl = object_storage->readObject(object_, read_settings, read_hint, file_size);
-
-        if ((!object_storage->supportsCache() || !read_settings.enable_filesystem_cache)
-            && read_settings.page_cache && read_settings.use_page_cache_for_disks_without_file_cache)
-        {
-            /// Can't wrap CachedOnDiskReadBufferFromFile in CachedInMemoryReadBufferFromFile because the
-            /// former doesn't support seeks.
-            auto cache_path_prefix = fmt::format("{}:", magic_enum::enum_name(object_storage->getType()));
-            const auto object_namespace = object_storage->getObjectsNamespace();
-            if (!object_namespace.empty())
-                cache_path_prefix += object_namespace + "/";
-
-            const auto cache_key = FileChunkAddress { .path = cache_path_prefix + object_.remote_path };
-
-            impl = std::make_unique<CachedInMemoryReadBufferFromFile>(
-                cache_key, read_settings.page_cache, std::move(impl), read_settings);
-        }
-        return impl;
+        return object_storage->readObject(object_, read_settings, read_hint, file_size);
     };
 
-    /// Avoid cache fragmentation by choosing bigger buffer size.
+    /// Avoid cache fragmentation by choosing a bigger buffer size.
+    /// But don't use it if the cache is used passively (only for reading if data is already cached, such as during merges).
     bool prefer_bigger_buffer_size = read_settings.filesystem_cache_prefer_bigger_buffer_size
+        && !read_settings.read_from_filesystem_cache_if_exists_otherwise_bypass_cache
         && object_storage->supportsCache()
         && read_settings.enable_filesystem_cache;
 
     size_t buffer_size = prefer_bigger_buffer_size
-        ? std::max<size_t>(settings.remote_fs_buffer_size, DBMS_DEFAULT_BUFFER_SIZE)
+        ? std::max<size_t>(settings.remote_fs_buffer_size, settings.prefetch_buffer_size)
         : settings.remote_fs_buffer_size;
 
     size_t total_objects_size = file_size ? *file_size : getTotalSize(storage_objects);
     if (total_objects_size)
         buffer_size = std::min(buffer_size, total_objects_size);
 
-    const bool use_async_buffer = read_settings.remote_fs_method == RemoteFSReadMethod::threadpool;
-    auto impl = std::make_unique<ReadBufferFromRemoteFSGather>(
+    std::unique_ptr<ReadBufferFromFileBase> impl;
+    impl = std::make_unique<ReadBufferFromRemoteFSGather>(
         std::move(read_buffer_creator),
         storage_objects,
         read_settings,
-        global_context->getFilesystemCacheLog(),
-        /* use_external_buffer */use_async_buffer,
-        /* buffer_size */use_async_buffer ? 0 : buffer_size);
+        use_external_buffer_for_gather,
+        /* buffer_size */use_external_buffer_for_gather ? 0 : buffer_size);
+
+    if (use_page_cache)
+    {
+        /// We identify the file by its first object, with the assumption that an object can't
+        /// belong to more than one file.
+        auto cache_path_prefix = fmt::format("{}:{}:", /*disk*/ name, magic_enum::enum_name(object_storage->getType()));
+        const auto object_namespace = object_storage->getObjectsNamespace();
+        if (!object_namespace.empty())
+            cache_path_prefix += object_namespace + "/";
+        const auto cache_key = PageCacheKey { .path = cache_path_prefix + storage_objects.at(0).remote_path };
+
+        impl = std::make_unique<CachedInMemoryReadBufferFromFile>(
+            cache_key, read_settings.page_cache, std::move(impl), read_settings);
+    }
 
     if (use_async_buffer)
     {
@@ -809,31 +838,6 @@ void DiskObjectStorage::applyNewSettings(
     remove_shared_recursive_file_limit = config.getUInt64(config_prefix + ".remove_shared_recursive_file_limit", DEFAULT_REMOVE_SHARED_RECURSIVE_FILE_LIMIT);
 
     IDisk::applyNewSettings(config, context_, config_prefix, disk_map);
-}
-
-void DiskObjectStorage::restoreMetadataIfNeeded(
-    const Poco::Util::AbstractConfiguration & config, const std::string & config_prefix, ContextPtr context)
-{
-    if (send_metadata)
-    {
-        metadata_helper->restore(config, config_prefix, context);
-
-        auto current_schema_version = DB::DiskObjectStorageRemoteMetadataRestoreHelper::readSchemaVersion(object_storage.get(), object_key_prefix);
-        if (current_schema_version < DiskObjectStorageRemoteMetadataRestoreHelper::RESTORABLE_SCHEMA_VERSION)
-            metadata_helper->migrateToRestorableSchema();
-
-        metadata_helper->findLastRevision();
-    }
-}
-
-void DiskObjectStorage::syncRevision(UInt64 revision)
-{
-    metadata_helper->syncRevision(revision);
-}
-
-UInt64 DiskObjectStorage::getRevision() const
-{
-    return metadata_helper->getRevision();
 }
 
 #if USE_AWS_S3
